% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{calcContrast}
\alias{calcContrast}
\title{Calculate Local Image Contrast}
\usage{
calcContrast(image, w_x, w_y, brushShape = "disc", validate = TRUE)
}
\arguments{
\item{image}{A numeric 2D matrix-like on which the contrast should be 
determined}

\item{w_x}{The size of the window in x-direction}

\item{w_y}{The size of the window in y-direction}

\item{brushShape}{A string indicating the shape of the window. Currently 
supported values are \code{"box", "disc"} and the default is "disc".}

\item{validate}{A boolean value indicating if the variables need to be 
validated or if this function is being called internally, i.e. the variables
have already been validated once. This is only used to marginally speed up 
internal calls to functions and has no bearing on the actual functionality 
of the method.}
}
\value{
A 2D matrix with the same dimensions as \code{image}, which shows 
the local contrast at each corresponding pixel of \code{image}.
}
\description{
Calculate the local contrast of an input image for every pixel 
of the image over a given window centered on that pixel.
}
\details{
The local contrast is calculated for every pixel of \code{image}. 
This means that a window is centered around a given pixel and the variance 
of the intensity values within this window are determined via 
Var = E[X^2] - (E[X])^2.

The \code{brushShape} indicates the shape of the window over which to 
calculate the variance. Depending on the symmetry of the objects being 
imaged, the window shape may have a significant impact on the quality of 
the projection.
}
\examples{
print(calcContrast)
}
\author{
Jan Sauer
}
\keyword{array}
