% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{intensityProjection}
\alias{intensityProjection}
\title{Intensity Projection}
\usage{
intensityProjection(imageStack, projType = "max")
}
\arguments{
\item{imageStack}{A numeric 3D array-like which should ne projected. The 
dimensions should be (spatial_1, spatial_2, numer_of_images)}

\item{projType}{A string indicating the type of projection. Defaults to 
"max" and can take on the following values: 
\code{"max", "min", "mean", "median", "sd", "sum"}}
}
\value{
A 2D matrix corresponding to the maximum intensity projection of 
\code{imageStack}
}
\description{
Perform an intensity projection of a stack of images. The type 
of projection depends on the \code{projType}
}
\details{
The \code{projType} determines the time of projection to be used:
\describe{
    \item{max}{Each pixel of the output image takes on the maximum value of 
    the z-stack underneath the corresponding pixel of the input image 
    stack.}
    \item{min}{Each pixel of the output image takes on the minimum value of 
    the z-stack underneath the corresponding pixel of the input image 
    stack.}
    \item{mean}{Each pixel of the output image takes on the mean value of 
    the z-stack underneath the corresponding pixel of the input image 
    stack.}
    \item{median}{Each pixel of the output image takes on the median value 
    of the z-stack underneath the corresponding pixel of the input image 
    stack.}
    \item{sd}{Each pixel of the output image takes on the standard 
    deviation of the values of the z-stack underneath the corresponding 
    pixel of the input image stack.}
    \item{sum}{Each pixel of the output image takes on the sum of the 
    values of the z-stack underneath the corresponding pixel of the input 
    image stack.}
}
}
\examples{
dat = array(c(1,1,2,2,1,2,3,1), dim = c(2,2,2))
proj = intensityProjection(dat, projType = "max")
print(proj)
}
\author{
Jan Sauer
}
\keyword{array}
