% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\docType{package}
\name{MaxContrastProjection}
\alias{MaxContrastProjection}
\alias{MaxContrastProjection-package}
\title{MaxContrastProjection: A package for performing z-projections of image 
stacks}
\description{
The package MaxContrastProjection provides functions to perform the 
common intensity projections (max, min, etc.) as well as a maximum 
contrast projection we introduce here.
}
\examples{
data(cells)
proj = contrastProjection(imageStack = cells, w_x = 15, smoothing = 5)
}
