% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{validateVariables}
\alias{validateVariables}
\title{Validate Function Arguments}
\usage{
validateVariables(imageStack, image, w_x, w_y, smoothing, brushShape, 
    projType, interpolation, contrastStack, indexMap, fix.gaussian.blur, 
    blur.size, return.all)
}
\arguments{
\item{imageStack}{A numeric 3D array-like}

\item{image}{A 2D array-like}

\item{w_x}{A numeric value}

\item{w_y}{A numeric value}

\item{smoothing}{An integer value}

\item{brushShape}{A string}

\item{projType}{A string}

\item{interpolation}{An integer value}

\item{contrastStack}{A numeric 3D array-like}

\item{indexMap}{A numeric 2D array-like. Must always be validated 
simultaneously with \code{imageStack}}

\item{fix.gaussian.blur}{A logical parameter, or one which can be coerced 
to a logical value}

\item{blur.size}{An integer value}

\item{return.all}{A logical parameter, or one which can be coerced to a 
logical value}
}
\value{
A boolean indicating that the function ran without error.
}
\description{
Validate the function arguments based on predefined rules. This
function has no application for users and serves only to ensure correct 
variables. It either returns TRUE if all variables conform to the rules or 
stops execution in case one of the conditions is not met.
}
\examples{
print(validateVariables)
}
\author{
Jan Sauer
}
\keyword{array}
