% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalVars.R
\docType{package}
\name{ZProjection}
\alias{ZProjection}
\alias{ZProjection-package}
\title{ZProjection: A package for performing z-projections of image stacks}
\description{
The package ZProjection provides functions to perform the common intensity projections (max, min, etc.) as well as a maximum contrast projection we introduce here.
}
\examples{
data(cells)
proj = contrastProjection(imageStack = cells, w_x = 15, smoothing = 5)
}

